Rename audioaccuracytest to wakewordtest
