+++ test/wake_word/README.md
